Rename Browser to ItemBrowser
'Browser' is just too generic a term in webdev.
